+++ NEWS.rst
@@ -1,6 +1,12 @@
 matlab_wrapper -- history of user-visible changes
 =================================================
 
+Changes in version NEXT
+-----------------------
+
++ Check if Python and MATLAB architectures match
+
+
 Changes in version 0.9.6
 ------------------------